pyproject.toml
[tool.poetry]
name = "dbms-backend"
version = "0.1.0"
description = ""
authors = ["Zzznorlax <legohulk27@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
fastapi = "^0.70.1"
SQLAlchemy = "^1.4.28"
requests = "^2.26.0"
uvicorn = "^0.16.0"
pydantic = {extras = ["dotenv"], version = "^1.8.2"}
alembic = "^1.7.5"
PyJWT = "^2.3.0"
python-multipart = "^0.0.5"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
